new: Synchronize maildir flags eagerly.
Because flag synchronization is stateless, it can be performed at any
time as long as it's guaranteed to be performed after any change to a
message's filename list.  Take advantage of this to synchronize tags
immediately after a filename is added or removed while the message is
still frozen.  With this change, filename removal is atomic and the
unsynchronized window for filename add is very short.