add signature-verification and decryption tests ("crypto")
This adds a new "crypto" test script to the test suite to test
PGP/MIME signature verification and message decryption.  Included here
is a test GNUPGHOME with a test secret key (passwordless).  Two test
messages are generated: one signed by this key, and another encrypted
to this key (also with signature).  The test then checks the output of
"--format=json --verify" and "--format=json --decrypt" respectively.