output signer user ID if validity is full or ultimate
This means that if the userid is included in the json output then it
has full or ultimate validity.  If the userid is not included that
means the userid has less than full validity, and we should be doing
something else to determine validity (probably of the From: name of
the message).