emacs: add support for PGP/MIME verification/decryption
This adds a new emacs configuration variable
"notmuch-show-process-pgpmime".  When this is set true, notmuch-query
will use the notmuch show --decrypt flag to decrypt encrypted messages
and/or calculate the sigstatus of signed messages.  If sigstatus is
available, notmuch-show will place a specially color-coded header at
the begining of the signed message.

Also included is the ability to switch decryption/verification on/off
on the fly, which is bound to M-RET in notmuch-search-mode.