emacs: change name of notmuch-process-crypto-mime variable
This changes the customization variable from "notmuch-process-pgpmime"
to "notmuch-process-crypto-mime".  There is indication that gmime will
soon support crypto operations on s/mime parts, with the same crypto
context that is currently being used for pgp/mime.  This means we
should be able to support s/mime crypto operations with the same code.
For this reason it's better to use a more general name for this
variable.

Sorry to the early adopters!