crypto: change customization variable name again
I sweat this is the last time.  I realized that it makes more sense to
scope the name space a bit, as I now realize there will be more crypto
function and customizations to come.  This commit precedes the
breaking out of the existing crypto functions in to their own
libraries.