Show addresses with a match to user's input anywhere.
The function completing-read only matches from the begginning, what
makes the search for an address when there is more than one match
unnecessarily harder unless the input typed matches exactly with the
beginning of all returned addresses.

Using Ido with flex-matching (called "fuzzy match" in other places)
has the effect of showing all addresses matching the user's input,
regardless of the position of the match. In addition to this, there
are other facilities for interactively refining the matches and for
selecting a match.