Enable decryption of parts in notmuch show --format=part.
As with --format=json, decryption is enabled with --decrypt.

This is a fairly straightforward patch that decrypts all
multipart/encrypted parts when retrieving individual parts with
format=part.  The mechanism is the same as how it is done in
format=json, so part numbering should always be consistent.  All parts
have to be decrypted internally to get the overall part-count correct.